Changed characters structures
Added Skill Class
Added CharacterUtils file and moving Item and Inventory class into it
Added properties to Character class
