+++ Character.py
@@ -23,124 +23,11 @@ from random import randint,choice
 from converter import *
 from DatabaseManager import *
 from BotTools import DBJDR
-
-class Item:
-    def __init__(self,name,descr,weight):
-        self.name = name
-        self.description = descr
-        self.weight = weight
-        self.ID = None
-
-    def __str__(self):
-        return self.name+" ("+self.weight+") : "+self.description
-
-    def create(self):
-        db = Database()
-        cur = db.call("createitem",name=self.name,descr=self.description,poids=self.weight)
-        if cur is None:
-            db.close(True)
-            raise DatabaseError("unable to create the item")
-        self.ID = cur.fetchone()[0]
-        db.close()
-
-    def load(self):
-        db = Database()
-        cur = db.execute("SELECT id_item FROM Items WHERE nom = %(nom)s AND description = %(descr)s AND weight = %(poids)s;",nom=self.name,descr=self.description,poids=self.weight)
-        if cur is None:
-            db.close(True)
-            raise DatabaseErrror("unable to load the item")
-        self.ID = cur.fetchone()[0]
-        db.close()
-
-    def delete(self):
-        if self.ID is None:
-            raise AttributeError("self.ID is not defined for this item, maybe it has not be loaded or created before")
-        db = Database()
-        db.call("deleteitem",item=self.ID)
-        db.close()
-
-    def find(name):
-        db = Database()
-        cur = db.execute("SELECT id_item,description,weight FROM Items WHERE nom = %(nom)s;",nom=name)
-        if cur is None:
-            db.close(True)
-            raise DatabaseError("unable to find the item")
-        it = cur.fetchone()
-        if it is None:
-            db.close(True)
-            return None
-        db.close()
-        item = Item(name,it[1],it[2])
-        item.ID = it[0]
-        return item
-    find = staticmethod(find)
-
-class Inventory:
-    def __init__(self,maxw=20):
-        self.items = {}
-        self.maxweight = 20
-        self.weight = 0
-        self.character = None
-        self.jdr = None
-        self.ID = None
-
-    def bind(self,char,jdr):
-        self.character = char
-        self.jdr = jdr
-
-    def loadfromdb(self,inventory_id):
-        db = Database()
-        cur = db.execute("SELECT nom,qte,description,weight FROM contient INNER JOIN items ON (contient.id_item = items.id_item) WHERE id_inventory = %(idinv)s;",idinv=inventory_id)
-        if cur is None:
-            db.close(True)
-            raise DatabaseException("unable to find the inventory")
-        self.items = {}
-        for i in cur:
-            it = Item(i[0],i[2],i[3])
-            it.load()
-            self.items[it] = i[1]
-        db.close()
-        self.ID = inventory_id
-        self._reload()
-
-    def _reload(self):
-        db = Database()
-        cur = db.execute("SELECT size_,size_max FROM inventaire WHERE id_inventory = %(idinv)s;",idinv=self.ID)
-        if cur is None:
-            db.close(True)
-            raise DatabaseErrror("unable to reload the inventory")
-        info = cur.fetchone()
-        db.close()
-        self.weight = info[0]
-        self.maxweight = info[1]
-
-    def __add__(self,it):
-        return self.additem(it,1)
-
-    def __iadd__(self,it):
-        return self.additem(it,1)
-
-    def __sub__(self,it):
-        return self.rmitem(it,1)
-
-    def __isub__(self,it):
-        return self.rmitem(it,1)
-
-    def additem(self,it,qte):
-        db = Database()
-        db.call("additem",dbkey=self.character.key,idserv=self.jdr.server,idchan=self.jdr.channel,itname=it.name,quantite=qte)
-        db.close()
-        self.loadfromdb(self.ID)
-        
-    def rmitem(self,it,qte):
-        db = Database()
-        db.call("removeitem",dbkey=self.character.key,idserv=self.jdr.server,idchan=self.jdr.channel,itname=it.name,quantite=qte)
-        db.close()
-        self.loadfromdb(self.ID)
+from CharacterUtils import *
 
 class Character:
     """Character class"""
-    def __init__(self,dic={"charkey":"","name":"","lore":"","PVm":1,"PMm":1,"force":50,"esprit":50,"charisme":50,"furtivite":50,"karma":0,"money":0,"stat":[0,0,0,0,0,0,0],"lp":0,"dp":0,"regenkarm":0.1,"mod":0,"armor":0,"RM":0,"PV":1,"PM":1,"default_mod":0,"default_karma":0,"intuition":3,"mentalhealth":100,"lvl":1,"linked":None,"inventory":Inventory(),"pet":{}}):
+    def __init__(self,dic={"charkey":"","name":"","lore":"","PVm":1,"PMm":1,"force":50,"esprit":50,"charisme":50,"furtivite":50,"karma":0,"money":0,"stat":[0,0,0,0,0,0,0],"lp":0,"dp":0,"regenkarm":0.1,"mod":0,"armor":0,"RM":0,"PV":1,"PM":1,"default_mod":0,"default_karma":0,"intuition":3,"mentalhealth":100,"lvl":1,"linked":None,"inventory":Inventory(),"pet":{},"skills":[],"dead":False,"classe":1}):
         self.key = dic["charkey"]
         self.name = dic["name"]
         self.lore = dic["lore"]
@@ -168,6 +55,9 @@ class Character:
         if self.linked.upper() == "NULL": self.linked = None
         self.inventory = dic["inventory"]
         self.pet = dic["pet"]
+        self.skills = dic["skills"]
+        self.dead = dic["dead"]
+        self.race,self.classe = retrieveCharacterOrigins(dic["classe"])
         self.jdr = None
         #mod 0 = offensiv / mod 1 = defensiv
